pyproject.toml
[project]
name = "VitalTrack"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Daniel Williams", email = "danielfwilliams@protonmail.com"},
]
dependencies = [
    "fastapi>=0.111.1",
    "python-dotenv>=1.0.1",
    "motor>=3.5.1",
    "pydantic>=2.8.2",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.3.2",
]
lint = [
    "pylint>=3.2.6",
    "black>=24.4.2",
]

[tool.pdm.scripts]
dev = """
    fastapi dev src/vitaltrack/main.py
"""
